Move secrets to datastore package
