+++ pkg/refactor/api/errors.go
@@ -95,7 +95,7 @@ func writeDataStoreError(w http.ResponseWriter, err error) {
 
 		return
 	}
-	if errors.Is(err, core.ErrSecretNotFound) {
+	if errors.Is(err, datastore.ErrSecretNotFound) {
 		writeError(w, &Error{
 			Code:    "resource_not_found",
 			Message: "requested resource is not found",
