+++ pkg/refactor/datastore/secrets.go
@@ -1,4 +1,4 @@
-package core
+package datastore
 
 import (
 	"context"
